Move nibble library to common
So we can use it outside jellyfish_merkle. See
https://github.com/libra/libra/pull/978 for mode details.

Made minor changes to `impl Arbitrary for Nibble`.
