+++ common/nibble/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "nibble"
+version = "0.1.0"
+authors = ["Libra Association <opensource@libra.org>"]
+license = "Apache-2.0"
+publish = false
+edition = "2018"
+
+[dependencies]
+proptest = "0.9.4"
+serde = { version = "1.0.101", features = ["derive"] }
