+++ common/nibble/src/lib.rs
@@ -1,0 +1,39 @@
+// Copyright (c) The Libra Core Contributors
+// SPDX-License-Identifier: Apache-2.0
+
+//! `Nibble` represents a four-bit unsigned integer.
+
+use proptest::prelude::*;
+use serde::{Deserialize, Serialize};
+use std::fmt;
+
+#[derive(Clone, Copy, Debug, Hash, Eq, PartialEq, Serialize, Deserialize)]
+pub struct Nibble(u8);
+
+impl From<u8> for Nibble {
+    fn from(nibble: u8) -> Self {
+        assert!(nibble < 16, "Nibble out of range: {}", nibble);
+        Self(nibble)
+    }
+}
+
+impl From<Nibble> for u8 {
+    fn from(nibble: Nibble) -> Self {
+        nibble.0
+    }
+}
+
+impl fmt::LowerHex for Nibble {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{:x}", self.0)
+    }
+}
+
+impl Arbitrary for Nibble {
+    type Parameters = ();
+    type Strategy = BoxedStrategy<Self>;
+
+    fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
+        (0..16u8).prop_map(|u| Self::from(u)).boxed()
+    }
+}
