+++ storage/jellyfish_merkle/Cargo.toml
@@ -13,10 +13,11 @@ num-derive = "0.2"
 num-traits = "0.2"
 proptest = "0.9.2"
 proptest-derive = "0.1.2"
+serde = { version = "1.0.89", features = ["derive"] }
 
 crypto = { path = "../../crypto/crypto" }
 failure = { path = "../../common/failure_ext", package = "failure_ext" }
-serde = { version = "1.0.89", features = ["derive"] }
+nibble = { path = "../../common/nibble" }
 types = { path = "../../types" }
 
 [dev-dependencies]
