+++ storage/jellyfish_merkle/src/iterator/mod.rs
@@ -10,12 +10,13 @@
 mod iterator_test;
 
 use crate::{
-    nibble::{Nibble, NibblePath},
+    nibble_path::NibblePath,
     node_type::{InternalNode, Node, NodeKey},
     TreeReader,
 };
 use crypto::HashValue;
 use failure::prelude::*;
+use nibble::Nibble;
 use types::{account_state_blob::AccountStateBlob, transaction::Version};
 
 /// `NodeVisitInfo` keeps track of the status of an internal node during the iteration process. It
